backend/pyproject.toml
[project]
name = "bundesfaq-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.1.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "jq>=1.10.0",
    "langchain==0.1.20",
    "langchain-community==0.0.38",
    "langchain-core==0.1.52",
    "langchain-experimental>=0.0.58",
    "langchain-groq>=0.1.5",
    "langchain-huggingface>=0.0.3",
    "langchain-openai>=0.1.7",
    "langgraph>=0.0.51",
    "matplotlib>=3.10.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "tiktoken>=0.11.0",
    "unstructured>=0.18.15",
    "uvicorn[standard]>=0.35.0",
    "wikipedia>=1.4.0",
    "tomli>=2.0.1",
]
